Add system to backup and generate a site
This runs a backup correctly and atomically on a minecraft server, and
then produces an index of all the backups uploaded to an AWS S3 bucket.
It relies on mcrcon (https://github.com/Tiiffi/mcrcon) in order to tell
the server to stop saving temporarily (to ensure atomicity of a save),
and the AWS CLI to upload the files.